Next release

See merge request deflect-dev/deflect!31

+++ orchestration/run_container/nginx.py
@@ -28,12 +28,20 @@ class Nginx(Container):
         self.container.exec_run("mv /etc/ssl/archive /etc/ssl/sites")
 
         # XXX specific to our internals
-        # with open(f"{path_to_output()}/{config_timestamp}/etc-ssl-uploaded.tar", "rb") as f:
-        #     self.container.put_archive("/etc/ssl-uploaded/", f.read())
+        try:
+            with open(f"{path_to_output()}/{config_timestamp}/etc-ssl-uploaded.tar", "rb") as f:
+                self.container.put_archive("/etc/ssl-uploaded/", f.read())
+        except FileNotFoundError:
+            self.logger.info("No etc-ssl-uploaded.tar found in output")
 
         # XXX note that sending this signal does not guarantee the new config is actually loaded.
         # the config might be invalid.
-        self.container.kill(signal="SIGHUP")
+        (exit_code, output) = self.container.exec_run("nginx -t")
+        if exit_code != 0:
+            self.logger.error(f"nginx config failed validation. output: {output}")
+            raise Exception("nginx config failed validation")
+        else:
+            self.container.kill(signal="SIGHUP")
 
         self.logger.info("installed new config + certs on nginx container")
